requirements.txt
Kurigram
TgCrypto
uvloop
python-dotenv
